Restructure HTTP API
