+++ src/smif/http_api/app.py
@@ -1,27 +1,11 @@
-"""Provide APP constant for the purposes of manually running the flask app
 
-For example, build the front end::
-        cd smif/app/
-        npm run build
-
-Or to rebuild js/css on change:
-        npm run watch
-
-Then run the server app in debug mode with environment variables
-        FLASK_APP=smif.http_api.app FLASK_DEBUG=1 flask run
-
-On Windows under some Flask versions, use this workaround for 'SyntaxError:
-Non-UTF-8 code starting with '\x90' in file flask.exe'
-(see https://github.com/pallets/flask/issues/2543):
-        FLASK_APP=smif.http_api.app FLASK_DEBUG=1 python -m flask run
-
-Or if backend debug mode is not needed, just use the smif CLI:
-        smif app -d ../sample_project
-"""
 import pkg_resources
 
+from flask import Flask
+from smif.controller import Scheduler
 from smif.data_layer import DatafileInterface
-from smif.http_api import create_app
+from smif.http_api.register import (register_api_endpoints,
+                                    register_error_handlers, register_routes)
 
 
 def get_data_interface():
@@ -32,8 +16,28 @@ def get_data_interface():
     )
 
 
-APP = create_app(
-    static_folder=pkg_resources.resource_filename('smif', 'app/dist'),
-    template_folder=pkg_resources.resource_filename('smif', 'app/dist'),
-    get_data_interface=get_data_interface
-)
+def get_scheduler():
+    """Return a controller.Scheduler
+    """
+    return Scheduler()
+
+
+def create_app(static_folder='static', template_folder='templates',
+               data_interface=get_data_interface(), scheduler=get_scheduler()):
+    """Create Flask app object
+    """
+    app = Flask(
+        __name__,
+        static_url_path='',
+        static_folder=static_folder,
+        template_folder=template_folder
+    )
+
+    app.config.data_interface = data_interface
+    app.config.scheduler = scheduler
+
+    register_routes(app)
+    register_api_endpoints(app)
+    register_error_handlers(app)
+
+    return app
